# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIX for JSSR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/member/login": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Member Login",
                "description": "Login Member Here",
                "operationId": "memberLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/member/logout": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Member Logout",
                "description": "Member Logout",
                "operationId": "MemberLogout",
                "responses": {
                    "200": {
                        "description": "Logout Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/member/refresh": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Refresh Token",
                "description": "Refresh Token",
                "operationId": "MemberRefreshToken",
                "responses": {
                    "200": {
                        "description": "Get Member Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/member/user-profile": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "Get Member Detail",
                "description": "Get Member Detail",
                "operationId": "GetMemberDetail",
                "responses": {
                    "200": {
                        "description": "Get Member Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "developer"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "developer@jssr.co.th"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "example": "01/01/2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Login",
                "description": "Login Admin Here",
                "operationId": "adminAuthLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Logout",
                "description": "Admin Logout",
                "operationId": "AdminLogout",
                "responses": {
                    "200": {
                        "description": "Logout Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/refresh": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Refresh Token",
                "description": "Refresh Token",
                "operationId": "adminRefreshToken",
                "responses": {
                    "200": {
                        "description": "Get Admin Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/user-profile": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Admin Detail",
                "description": "Get Admin Detail",
                "operationId": "GetAdminDetail",
                "responses": {
                    "200": {
                        "description": "Get User Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "developer"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "developer@jssr.co.th"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "example": "01/01/2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}